Add entrypoint for aws s3 syncing
We use the s3 sync action from the aws cli to sync the repo with the
bucket in S3. The --delete option means that we clean up any files which
are removed from the theme repo in S3.
The three required keys are what the user of the action will set on the
theme repo as supplied by an Easol member of staff.